\name{amavg
\alias{amavg}

\title{
Simultaneous inference of model-averaged derived parameters
}
\description{
something to be changed
}
\usage{
amavg(modelList, modelWeights, expressions, interval=c("Buckland", "Asymptotic"),level=0.95,
                 distribution="normal",seType="mod")
}

\arguments{
  \item{modelList}{}
  \item{modelWeights}{}
  \item{expressions}{}
  \item{interval}{}
  \item{level}{}
  \item{distribution}{}
  \item{seType}{}
  }

\details{
  }.
}

\value{
  A list.
}

\references{
  Jensen, S. M., & Ritz, C. (2015). 
  Simultaneous inference for model averaging of derived parameters. 
  \emph{Risk Analysis} \bold{35(1)}, 68--76.
  
}

\author{Signe M. Jensen}

%\note{}

%\seealso{}

\examples{


}
\keyword{Model-averaging}

\concept{Model-averaging}