\name{MEDFinding}
\alias{MEDFinding}

\title{
a title
}
\description{
something to be changed
}
\usage{
MEDFinding(object, data, type="normal", useModels, delta,pocMethods,sMethod="AIC", userD = NA, 
                     margin=0.0001, userContr=NA, adjusted=TRUE, Quantile=FALSE)
}

\arguments{
  \item{object}{}
  \item{data}{}
  \item{type}{}
  \item{useModels}{}
  \item{delta}{}
  \item{pocMethods}{}
  \item{sMethod}{}
  \item{userD}{}
  \item{margin}{}
  \item{userContr}{}
  \item{adjusted}{}
  \item{Quantile}{}
}

\details{
  }.
}

\value{
  A list.
}

\references{
  
  
}

\author{Signe M. Jensen}

%\note{}

%\seealso{}

\examples{


}
\keyword{Dose-finding}
\keyword{Model-averaging}
\keyword{Iterative dose-finding}

\concept{Dose-finding}