\name{POC}
\alias{POC}

\title{
Proof of concept
}
\description{
something to be changed
}
\usage{
POC(object, data, type = "normal", pocMethods=c("linear","Dunnett","DRModels","MCP-MOD","user"), userContr=NA, useModels=NA)
}

\arguments{
  \item{object}{}
  \item{data}{}
  \item{type}{}
  \item{pocMethods}{}
  \item{useModels}{}
  \item{userContr}{}
  }

\details{
 
}

\value{
  A list.
}

\references{
  
  
}

\author{Signe M. Jensen}

%\note{}

%\seealso{}

\examples{


}
\keyword{Proof of concept}
\keyword{Contrast testing}

\concept{Dose-finding}